30-9-2022    from home

+++ lib/data/model/signModel/Ratings.dart
@@ -1,13 +1,14 @@
 class Ratings {
   Ratings({
-      int? drive, 
-      int? car, 
-      int? guide, 
-      int? attitude, 
-      int? driveAverage, 
-      int? carAverage, 
-      int? guideAverage, 
-      int? attitudeAverage,}){
+    int? drive,
+    int? car,
+    int? guide,
+    int? attitude,
+    int? driveAverage,
+    int? carAverage,
+    int? guideAverage,
+    double? attitudeAverage,
+  }) {
     _drive = drive;
     _car = car;
     _guide = guide;
@@ -16,7 +17,7 @@ class Ratings {
     _carAverage = carAverage;
     _guideAverage = guideAverage;
     _attitudeAverage = attitudeAverage;
-}
+  }
 
   Ratings.fromJson(dynamic json) {
     _drive = json['drive'];
@@ -28,6 +29,7 @@ class Ratings {
     _guideAverage = json['guideAverage'];
     _attitudeAverage = json['attitudeAverage'];
   }
+
   int? _drive;
   int? _car;
   int? _guide;
@@ -35,16 +37,23 @@ class Ratings {
   int? _driveAverage;
   int? _carAverage;
   int? _guideAverage;
-  int? _attitudeAverage;
+  double? _attitudeAverage;
 
   int? get drive => _drive;
+
   int? get car => _car;
+
   int? get guide => _guide;
+
   int? get attitude => _attitude;
+
   int? get driveAverage => _driveAverage;
+
   int? get carAverage => _carAverage;
+
   int? get guideAverage => _guideAverage;
-  int? get attitudeAverage => _attitudeAverage;
+
+  double? get attitudeAverage => _attitudeAverage;
 
   Map<String, dynamic> toJson() {
     final map = <String, dynamic>{};
@@ -58,5 +67,4 @@ class Ratings {
     map['attitudeAverage'] = _attitudeAverage;
     return map;
   }
-
 }
